feat: add next-auth connection
