+++ web-client/package.json
@@ -19,6 +19,7 @@
     "cookies-next": "^5.0.2",
     "lucide-react": "^0.471.0",
     "next": "15.1.4",
+    "next-auth": "5.0.0-beta.25",
     "next-intl": "^3.26.3",
     "next-themes": "^0.4.4",
     "react": "^19.0.0",
