+++ web-client/src/lib/auth/actions.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { signIn } from "./";
+
+export async function signInWithCredentials(credentials: any) {
+  return await signIn("credentials", credentials);
+}
